Implemented: Remaining content type tests
If a draft of a content type is created it will receive a _new_ ID
(autoincrement). If the draft is published all content and groups will
be migrated to use the new type id. The old version of the type will be
removed.

I wonder if this will cause problems with external data, which might reference
a type ID…