[WIP] A take on a somewhat genric solution to the cache clearing problem
For discussions with JV and PS to unblock GL on ContentType deletion.

Allows Slots to extend CollectorSlot class to gain access to be able to collect random data needed for a later emit.
We might want to solve this problem in a completly different way so made sure to mark it all as internal and deprecated.

But, on positive side:
- allows slot to collect affected nodes before deletion, and delete cache using that data after commit
- make sure slots or signals  does not need to keep track of transaction id's or something like that
- generic

Negative:
- generic, unless slots themselves uses types to make sure they only take collected data relevant to
- collectors tied to slots, might be decoupled: think several disconnected collectors to generate affected location ids for use by DeleteContentSlot

But probably good enough for intenral use for now.